LayoutTests/platform/chromium/fast/js/object-prototype-properties-expected.txt

This is a test case for bug 64678.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.prototype.toString.call(undefined) is "[object Undefined]"
PASS Object.prototype.toString.call(null) is "[object Null]"
PASS Object.prototype.toLocaleString.call(undefined) threw exception TypeError: Object.prototype.toLocaleString called on null or undefined.
PASS Object.prototype.valueOf.call(undefined) threw exception TypeError: Cannot convert null to object.
PASS Object.prototype.hasOwnProperty.call(undefined, 'hasOwnProperty') threw exception TypeError: Cannot convert null to object.
PASS Object.prototype.propertyIsEnumerable.call(undefined, 'propertyIsEnumerable') threw exception TypeError: Cannot convert null to object.
PASS Object.prototype.isPrototypeOf.call(undefined, this) threw exception TypeError: Object.prototype.isPrototypeOf called on null or undefined.
PASS successfullyParsed is true

TEST COMPLETE
